remove pdf renderers

+++ src/orders/orders.controller.ts
@@ -6,18 +6,11 @@ import { TiendaNubeEventDto } from 'src/infra/tiendanube/EventDto';
 export class OrdersController {
   constructor(private readonly ordersService: OrdersService) {}
 
-  @Get()
-  testOrder() {
-    // const tiendanubeClient = new TiendaNubeClient();
-    return;
-  }
-
   @Post('')
   receiveEvent(@Body() body: TiendaNubeEventDto) {
     if (['order/created', 'order/updated'].includes(body.event))
       return this.ordersService.upsert(body);
     if (body.event === 'order/packed') return null;
-    if (body.event === 'order/deleted')
-      return this.ordersService.delete(body);
+    if (body.event === 'order/deleted') return this.ordersService.delete(body);
   }
 }
